upgrade/asr-worker/requirements.txt
faster-whisper==1.1.1
torch==2.6.0
redis==5.0.8
mysql-connector-python==9.1.0
